Refactor code, separate static and instance classes.
The static code with the MediaWiki callback functions was cleanly separated
from the actual business code of the extension, and a PubmedParser namespace
was introduced. Static functions are now assembled in the
PubmedParser\Extension class; the main extension business logic resides in
PubmedParser\Core, and the XML parsing of articles is in
PubmedParser\Article.

* Change: The extension requires PHP 5.3.0 or later.